requirements.txt
agno
streamlit
google-generativeai
python-dotenv
psycopg2-binary
openai
requests
qdrant-client
